# api/swagger/swagger_output.json
{
  "swagger": "2.0",
  "info": {
    "version": "1.0.0",
    "title": "Projeto EcosRev 🍀",
    "description": "➡️Documentação gerada automaticamente pelo módulo <a href='https://github.com/davibaltar/swagger-autogen' target='_blank'>swagger-autogen</a>."
  },
  "host": "localhost:3000",
  "basePath": "",
  "schemes": [
    "http"
  ],
  "securityDefinitions": {
    "apiKeyAuth": {
      "type": "apiKey",
      "in": "header",
      "name": "access-token",
      "description": "Token de Acesso gerado após o login"
    }
  },
  "consumes": [
    "application/json"
  ],
  "produces": [
    "application/json"
  ],
  "paths": {
...
    "/api/beneficio/": {
      "get": {
        "tags": [
          "Benefícios"
        ],
        "summary": "GET recebendo todos os benefícios",
        "description": "Função chamada para executar o GET com todos os benefícios a ser exibido",
        "parameters": [
          {
            "name": "limit",
            "in": "query",
            "type": "string"
          },
          {
            "name": "skip",
            "in": "query",
            "type": "string"
          },
          {
            "name": "order",
            "in": "query",
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          },
          "500": {
            "description": "Erro Interno do Servidor"
          }
        },
        "security": [
          {
            "apiKeyAuth": []
          }
        ]
      },
      "post": {
        "tags": [
          "Benefícios"
        ],
        "summary": "POST para cadastrar um novo benefício",
        "description": "Função chamada para executar o POST adicionando um novo benefício",
        "responses": {
          "201": {
            "description": "Criado"
          },
          "400": {
            "description": "Solicitação Inválida"
          },
          "500": {
            "description": "Erro Interno do Servidor"
          }
        },
        "security": [
          {
            "apiKeyAuth": []
          }
        ]
      },
      "put": {
        "tags": [
          "Benefícios"
        ],
        "summary": "PUT recebendo um único beneficio pelo ID a ser modificado",
        "description": "Função chamada para executar o PUT com um único benefício a ser modificado pelo seu ID",
        "parameters": [
          {
            "name": "body",
            "in": "body",
            "schema": {
              "type": "object",
              "properties": {
                "_id": {
                  "example": "any"
                }
              }
            }
          }
        ],
        "responses": {
          "202": {
            "description": "Aceito"
          },
          "400": {
            "description": "Solicitação Inválida"
          },
          "500": {
            "description": "Erro Interno do Servidor"
          }
        },
        "security": [
          {
            "apiKeyAuth": []
          }
        ]
      }
    },
    "/api/beneficio/gt": {
      "get": {
        "tags": [
          "Benefícios"
        ],
        "summary": "GET recebendo todos os benefícios com filtros",
        "description": "Função chamada para executar o GET com todos os benefícios a ser exibido utilizando um filtro",
        "parameters": [
          {
            "name": "limit",
            "in": "query",
            "type": "string"
          },
          {
            "name": "skip",
            "in": "query",
            "type": "string"
          },
          {
            "name": "order",
            "in": "query",
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          },
          "500": {
            "description": "Erro Interno do Servidor"
          }
        },
        "security": [
          {
            "apiKeyAuth": []
          }
        ]
      }
    },
    "/api/beneficio/id/{id}": {
      "get": {
        "tags": [
          "Benefícios"
        ],
        "summary": "GET recebendo um benefício pelo ID",
        "description": "Função chamada para executar o GET exibindo apenas um único benefício pelo seu ID",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          },
          "500": {
            "description": "Erro Interno do Servidor"
          }
        },
        "security": [
          {
            "apiKeyAuth": []
          }
        ]
      }
    },
    "/api/beneficio/nome/{filtro}": {
      "get": {
        "tags": [
          "Benefícios"
        ],
        "summary": "GET recebendo todos os benefícios com filtros",
        "description": "Função chamada para executar o GET com todos os benefícios a ser exibido utilizando um filtro",
        "parameters": [
          {
            "name": "filtro",
            "in": "path",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          },
          "500": {
            "description": "Erro Interno do Servidor"
          }
        },
        "security": [
          {
            "apiKeyAuth": []
          }
        ]
      }
    },
    "/api/beneficio/{id}": {
      "delete": {
        "tags": [
          "Benefícios"
        ],
        "summary": "DELETE recebendo um benefício pelo ID",
        "description": "Função chamada para executar o DELETE de apenas um benefício pelo seu ID",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          },
          "404": {
            "description": "Não Encontrado"
          }
        },
        "security": [
          {
            "apiKeyAuth": []
          }
        ]
      }
    },
    "/api/beneficio/resgate": {
      "put": {
        "tags": [
          "Benefícios"
        ],
        "summary": "PUT recebendo o ID do benefício no qual será alterado pelo resgate",
        "description": "Função chamada para executar o PUT com um benefício específico a ser modificado pelo resgate do mesmo",
        "parameters": [
          {
            "name": "body",
            "in": "body",
            "schema": {
              "type": "object",
              "properties": {
                "_id": {
                  "example": "any"
                },
                "quantidade": {
                  "example": "any"
                }
              }
            }
          }
        ],
        "responses": {
          "202": {
            "description": "Aceito"
          },
          "400": {
            "description": "Solicitação Inválida"
          },
          "500": {
            "description": "Erro Interno do Servidor"
          }
        },
        "security": [
          {
            "apiKeyAuth": []
          }
        ]
      }
    },
...
  },
...
}